+++ libdl/libdl.h
@@ -6,7 +6,7 @@
 /*   By: jkoskela <jkoskela@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/28 02:46:16 by jkoskela          #+#    #+#             */
-/*   Updated: 2020/09/09 19:32:30 by jkoskela         ###   ########.fr       */
+/*   Updated: 2020/09/10 18:03:29 by jkoskela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,6 +35,6 @@ void		dl_print_b(t_dlist *node);
 void		dl_del_node(t_dlist **ref, t_dlist *del);
 void		dl_del_node_at(t_dlist **ref, size_t pos);
 void		dl_del_last(t_dlist **ref);
-void		dl_rotate(t_dlist **head_ref, int N);
+void		dl_rotate(t_dlist **head_ref, size_t N);
 
 #endif
